+++ app/components/index.ts
@@ -1,1 +1,11 @@
+// exporting as default
 export { default as Skeleton } from "./Skeleton";
+export { default as AddToCart } from "./AddToCart";
+export { default as Link } from "./Link";
+export { default as ProductCard } from "./ProductCard";
+
+// exporting as explicit name
+export { CallOutError } from "./CallOutError";
+export { ErrorInlineMessage } from "./ErrorInlineMessage";
+export { IssueStatusBadge } from "./IssueStatusBadge";
+export { Spinner } from "./Spinner";
