+++ app/issues/[id]/page.tsx
@@ -1,4 +1,4 @@
-import { IssueStatusBadge } from "@/app/components/IssueStatusBadge";
+import { IssueStatusBadge } from "@/app/components";
 import prisma from "@/prisma/client";
 import { Card, Flex, Heading, Text } from "@radix-ui/themes";
 import { Link, Button } from "@radix-ui/themes";
